Sample Generated Semantics
Isabelle will let us generate code for the inductive semantics but it
is not the typical function application, it produces a function that will
return a bool indicating if the second parameter is a result of the first.

We can try and enumerate the solution with a constraint solver but not all
languages (i.e. Scala) have such tools built in.