Canonicalization 1 sorry left
All the sorrys have been crammed into one lemma which is potentially
incorrect. The remaining lemma is that if adding a completely
independant node to the graph doesn't change the value a node evaluates to.

I've managed to get it down to the last 20 node kinds but I suspect
from here on it might be individual proofs.