Evaluation Value Remains Same
Added a new lemma 'stay_same' (name WIP) that states that
if no nodes that a nid uses (including the nid itself) changes
(where changes means that the IRNode type is different) between
graphs, then the evaluated value will be the same in both graphs.

Used this lemma to finish the eval_independent lemma.

Currently stuck on trying to prove nodes which use 'nid' in their
premises, for some reason the induction tactic refuses to consider
the nid identifier in the premises as equal to nid in the conclusion.
Configuring IRNode's to hold the nid does resolve this issue but
needlessly duplicates the state.