InvokeNode and MethodCallTarget
Added a very simple semantics for the InvokeNode and MethodCallTarget
to later replace the CallNode in order to match Graal more closely
and allow programs to be translated fully automatically.

Currently this setup does only handles recursive method calls.

To handle calls to other methods we will need:
1. To match the signature in the MethodCallTargetNode with the name
of a method.
2. An additional input to step_top that maps method signatures to
their IRGraph.