Redefined a floating node
Redefined a floating node to be any node that doesn't have a control flow.

Defined having a control flow as having successors edges or being one of the end nodes.

Attempted to prove that if all inputs to a node eval to an intval and the node is floating
then it evals to an intval... failed so far.