Simplify Semantics & Expand Well-formedness
Cleaned up the semantic rules so they are presented better in papers.

Expanded the wellformedness definition to include some invariants I suspect
to be true but will have to get Sadra to check. I don't think these
invariants need to be included in the paper as it doesn't help us yet.