Notation; inputs/succ; general heap
Added notation commands to write the kind function out as g<<nid>>
to try to save some space and make definitions clearer.
Additionally added NoNode as empty which may be worth removing.

Changed the name of inp to inputs.
Changed the return of inputs and succ to sets as it is by
far the most common use case.

Changed DynamicHeap to a generic ('a, 'b) -> Value map.